TodoList/requirements.txt
django
djangorestframework
pyjwt
python-dotenv
psycopg2